Update k3ng_keyer.ino
Change the default up/down speed change direction using the paddle press.
Modify the default code so that the left paddle decreases the speed/sidetone and the right paddle increases the speed/sidetone. Could be more intuitive since 'knobs to the right' implies more...
The new option OPTION_SWAP_PADDLE_PARAMETER_CHANGE_DIRECTION controls whether this swap is in effect or not. Unless the option is defined, then the default paddle direction will be implemented.
There is no increase in the compiled code size, since only one piece of code is included in the final compiled code, as was the case before adding this option.