Update k3ng_keyer.ino
Previously, when exiting Command Mode the screen was cleared and left blank. The display would then only be restored if you entered a character from the paddles or pressed a memory button.
This code always restores the display as you exit Command Mode. It does this simply by rewriting the saved screen buffer arrays to the display.
Only relevant if FEATURE_DISPLAY is defined.
The code for the for loop was plagiarised from the service_display() function.