Hack to support 00-09 as a version
Patch the admin 09 command to return a fake version instead of a
different arbitrary value, because Skookum Logger depends on it.

Admin command 9 returns a calibration value on WK1, 0 on WK2, and the
major version on WK3 (according to docs)

Skookum Logger expects a real version to come back on a WK2 device since
that's what the author's device does.

https://groups.io/g/radioartisan/topic/100917839#18017
https://groups.io/g/SkookumLogger/topic/100998043#51